Split up the API_EXCLUDES_CLIENTS and API_EXCLUDES_DOMAINS filters
