+++ src/routes/stats/history/filters/mod.rs
@@ -12,16 +12,17 @@ mod blocked;
 mod client;
 mod dnssec;
 mod domain;
+mod exclude_clients;
+mod exclude_domains;
 mod private;
 mod query_type;
 mod reply;
 mod setup_vars;
-mod setup_vars_exclude;
 mod status;
 mod time;
 mod upstream;
 
 pub use self::{
-    blocked::*, client::*, dnssec::*, domain::*, private::*, query_type::*, reply::*,
-    setup_vars::*, setup_vars_exclude::*, status::*, time::*, upstream::*
+    blocked::*, client::*, dnssec::*, domain::*, exclude_clients::*, exclude_domains::*,
+    private::*, query_type::*, reply::*, setup_vars::*, status::*, time::*, upstream::*
 };
